Use serializable database transactions for query
